ppod smaller demos

+++ code_examples/train_obstacle_tower/ppod/record_demo.py
@@ -114,11 +114,13 @@ def record():
                     filename = os.path.join(
                         args.demos_dir, "obstacletower_demo_{}".format(num))
 
+                import ipdb; ipdb.set_trace()  # TODO. how to use less memory?
+
                 np.savez(
                     filename,
-                    Observation=np.array(np.stack(obs_rollouts).astype(np.float32)).squeeze(1),
-                    Reward=np.array(np.stack(rews_rollouts).astype(np.float32)).squeeze(1),
-                    Action=np.expand_dims(np.array(np.stack(actions_rollouts).astype(np.float32)), axis=1)
+                    Observation=np.array(np.stack(obs_rollouts).astype(np.uint8)).squeeze(1),
+                    Reward=np.array(np.stack(rews_rollouts).astype(np.float16)).squeeze(1),
+                    Action=np.expand_dims(np.array(np.stack(actions_rollouts).astype(np.int8)), axis=1)
                 )
 
                 sys.exit()
